# Extra arguments to control behaviour of {rstpm2} fitting function