@entjos 2024-06-18
Hi @ellessenne!

Thank you for this pull request! Adding the control and `...` argument looks like a good addition to the `JointFPM` function.

I think the test just failed because the new arguments were not added to the `JointFPM` help file. I just added these.

Thanks for your suggestions 🙂

Best,
Joshua